Refactor dependency injection: rename app_dependency_injection.dart to app_di.dart and update related imports in home_screen.dart and main.dart
