+++ lib/home_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:ugaoo/core/environment/environment_config_manager.dart';
-import 'package:ugaoo/core/global_di/app_dependency_injection.dart';
+import 'package:ugaoo/core/global_di/app_di.dart';
 
 class HomeScreen extends StatelessWidget {
   const HomeScreen({super.key});
